Implement methods for reviewer user name and number of qa reports

+++ dataland-qa-service/src/main/kotlin/org/dataland/datalandqaservice/model/QaReviewResponse.kt
@@ -53,7 +53,7 @@ data class QaReviewResponse(
         description = BackendOpenApiDescriptionsAndExamples.NUMBER_QA_REPORTS_DESCRIPTION,
         example = BackendOpenApiDescriptionsAndExamples.NUMBER_QA_REPORTS_EXAMPLE,
     )
-    var numberQaReports: Int,
+    var numberQaReports: Long,
     @field:Schema(
         description = BackendOpenApiDescriptionsAndExamples.COMMENT_DESCRIPTION,
         example = BackendOpenApiDescriptionsAndExamples.COMMENT_EXAMPLE,
